refactoring towards OOPS, Mosh

+++ src/com/codewithmosh/MortgageCalculator/Console.java
@@ -4,21 +4,21 @@ import java.util.Scanner;
 
 public class Console {
 
-        private static final Scanner scanner = new Scanner(System.in);
+    private static final Scanner scanner = new Scanner(System.in);
 
-        public static double readNumber(String prompt) {
-            return scanner.nextDouble();
-        }
+    public static double readNumber(String prompt) {
+        return scanner.nextDouble();
+    }
 
-        public static double readNumber(String prompt, double min, double max) {
+    public static double readNumber(String prompt, double min, double max) {
         double value;
         while (true) {
             System.out.print(prompt);
-            value = scanner.nextFloat();
+            value = scanner.nextDouble();
             if (value >= min && value <= max)
                 break;
             System.out.println("Enter a value between " + min + " and " + max);
         }
         return value;
-        }
+    }
 }
